// src/types/env.ts
export type AppEnvType = {
    APP_ENV: string;

    APP_PROTOCOL: string;
    APP_HOST: string;

    PLANNER_API_PROTOCOL: string;
    PLANNER_API_HOST: string;
    PLANNER_API_ENDPOINT: string;

    CLOUD_STORAGE_POROTO_PLACE_IMAGES: string;
    CLOUD_STORAGE_IMAGE_BUCKET_NAME: string;
    CLOUD_STORAGE_IMAGE_BUCKET_PROTOCOL: string;
    CLOUD_STORAGE_IMAGE_BUCKET_HOST: string;

    IMAGE_OPTIMIZATION_API_PROTOCOL: string;
    IMAGE_OPTIMIZATION_API_HOST: string;
};